backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
SQLAlchemy==2.0.37
databases==0.9.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.7.1
email_validator==2.2.0
pydantic==2.10.6
passlib==1.7.4
python-jose==3.3.0
gunicorn==23.0.0
requests==2.32.3